requirements.txt
# PDF处理
PyMuPDF==1.23.8
pdf2image==1.16.3
Pillow==10.1.0

# LLM API客户端
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.3.2

# 实用工具
python-dotenv==1.0.0
tqdm==4.66.1
requests==2.31.0

# 输出格式
markdown==3.5.2
